updated to account for bytes instead of ints, added EOF instruction

+++ src/main/java/com/blackaby/Backend/Emulation/CPU/DuckCPU.java
@@ -38,7 +38,7 @@ public class DuckCPU {
     /**
      * This method queues an instruction
      */
-    public void queueInstruction(int opcode, int... values) {
+    public void queueInstruction(byte opcode, int... values) {
         InstructionType type = InstructionType.fromOpcode(opcode);
         switch (type) {
             case DEBUG_DISPLAY:
